requirements.txt
arro3-core==0.5.1
certifi==2025.8.3
charset-normalizer==3.4.2
deltalake==1.1.3
Deprecated==1.2.18
idna==3.10
numpy==2.3.2
pandas==2.3.1
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.2
